separate cohorts tests from manifests tests
