+++ tests/manifests_test.py
@@ -1,81 +1,6 @@
-import pytest
-import requests
 import json as json_utils
 import random
 from manifestservice import manifests
-
-from manifestservice.api import create_app
-
-mocks = {}
-
-
-@pytest.fixture
-def app(mocker):
-    test_user = {
-        "context": {
-            "user": {
-                "policies": [
-                    "data_upload",
-                    "programs.test-read-storage",
-                    "programs.test-read",
-                ],
-                "google": {"proxy_group": None},
-                "is_admin": True,
-                "name": "example@uchicago.edu",
-                "projects": {
-                    "test": [
-                        "read-storage",
-                        "read",
-                        "create",
-                        "write-storage",
-                        "upload",
-                        "update",
-                        "delete",
-                    ]
-                },
-            }
-        },
-        "aud": ["data", "user", "fence", "openid"],
-        "sub": "18",
-    }
-
-    mocks["current_token"] = mocker.patch(
-        "manifestservice.manifests.current_token", return_value=test_user
-    )
-
-    mocks["_authenticate_user"] = mocker.patch(
-        "manifestservice.manifests._authenticate_user", return_value=(None, 200)
-    )
-
-    mocks["_list_files_in_bucket"] = mocker.patch(
-        "manifestservice.manifests._list_files_in_bucket",
-        return_value=(
-            {
-                "manifests": [
-                    {"filename": "manifest-a-b-c.json"},
-                ],
-                "cohorts": [{"filename": "18e32c12-a053-4ac5-90a5-f01f70b5c2be"}],
-            },
-            True,
-        ),
-    )
-
-    mocks["_add_manifest_to_bucket"] = mocker.patch(
-        "manifestservice.manifests._add_manifest_to_bucket",
-        return_value=("manifest-xxx.json", True),
-    )
-
-    mocks["_get_file_contents"] = mocker.patch(
-        "manifestservice.manifests._get_file_contents", return_value=""
-    )
-
-    mocks["_add_GUID_to_bucket"] = mocker.patch(
-        "manifestservice.manifests._add_GUID_to_bucket",
-        return_value=("a-guid-value", True),
-    )
-
-    app = create_app()
-    return app
 
 
 def test_generate_unique_manifest_filename_basic_date_generation():
@@ -154,7 +79,7 @@ def test_is_valid_manifest():
     assert is_valid is True
 
 
-def test_POST_handles_invalid_json(client):
+def test_POST_handles_invalid_json(client, mocks):
     """
     Test that we get a 400 if flask.request.json is not filled in.
     """
@@ -162,7 +87,7 @@ def test_POST_handles_invalid_json(client):
     assert r.status_code == 400
 
 
-def test_POST_handles_invalid_manifest_keys(client):
+def test_POST_handles_invalid_manifest_keys(client, mocks):
     """
     Test that we get a 400 if the manifest is missing the required key -- object_id.
     """
@@ -176,7 +101,7 @@ def test_POST_handles_invalid_manifest_keys(client):
     assert r.status_code == 400
 
 
-def test_POST_successful_manifest_upload(client):
+def test_POST_successful_manifest_upload(client, mocks):
     """
     Test the full user pathway: a manifest is created, listed, and then downloaded.
     Unfortunately, we cannot verify here that the manifest is present in the listed files,
@@ -226,51 +151,3 @@ def test_POST_successful_manifest_upload(client):
     assert mocks["_add_manifest_to_bucket"].call_count == 1
     assert mocks["_list_files_in_bucket"].call_count == 1
     assert mocks["_get_file_contents"].call_count == 1
-
-
-def test_POST_successful_GUID_add(client):
-    """
-        Test the Export PFB to Workspace pathway: a cohort is added to the bucket.
-    Note that because s3 is being mocked, only an integration test can properly
-    verify file creation.
-    """
-
-    test_guid = "5183a350-9d56-4084-8a03-6471cafeb7fe"
-    post_body = {"guid": test_guid}
-
-    headers = {"Content-Type": "application/json", "Accept": "application/json"}
-    r = client.post("/cohorts", data=json_utils.dumps(post_body), headers=headers)
-
-    assert r.status_code == 200
-    assert mocks["_authenticate_user"].call_count == 1
-    assert mocks["_get_file_contents"].call_count == 0
-    assert mocks["_add_manifest_to_bucket"].call_count == 0
-    assert mocks["_add_GUID_to_bucket"].call_count == 1
-
-    json = r.json
-    new_guid = json["filename"]
-
-    assert new_guid is not None
-    assert type(new_guid) is str
-
-
-def test_GET_cohorts(client):
-    """
-    Test GET /cohorts
-    """
-
-    headers = {"Content-Type": "application/json", "Accept": "application/json"}
-    r = client.get("/cohorts", headers=headers)
-
-    assert r.status_code == 200
-    assert mocks["_authenticate_user"].call_count == 1
-    assert mocks["_get_file_contents"].call_count == 0
-    assert mocks["_add_manifest_to_bucket"].call_count == 0
-    assert mocks["_add_GUID_to_bucket"].call_count == 0
-    assert mocks["_list_files_in_bucket"].call_count == 1
-
-    json = r.json
-    cohorts_returned = json["cohorts"]
-    assert len(cohorts_returned) == 1
-    # From the s3 mock
-    assert cohorts_returned[0]["filename"] == "18e32c12-a053-4ac5-90a5-f01f70b5c2be"
